+++ src/plugins/media/index.ts
@@ -1,8 +1,11 @@
 import { BehaviorSubject, Subject } from '@reactivex/rxjs';
 import * as uuid from "uuid";
+import { viwiLogger } from "../../log";
+import * as stupid from 'stupid-player'; // a cross platform media player
 
 import { Service, Resource, Element, ResourceUpdate, StatusCode, ElementResponse, CollectionResponse } from "../viwiPlugin";
 import { RendererObject, CollectionObject, ItemObject } from "./schema";
+import { Service as Medialibrary, Tracks} from "../medialibrary";
 
 class Media extends Service {
   constructor() {
@@ -14,25 +17,31 @@ class Media extends Service {
 }
 
 interface RendererElement extends Element {
-  data: RendererObject
+  data: RendererObject;
+  player?: any
 }
 
 class Renderers implements Resource {
+
+  static netfluxRendererId = "d6ebfd90-d2c1-11e6-9376-df943f51f0d8";//uuid.v1();  // FIXED for now
+  static stdpRendererId = "deadbeef-d2c1-11e6-9376-beefdead";//uuid.v1();  // FIXED for now
+
   private _name:string;
   private _renderers:BehaviorSubject<RendererElement>[] = [];
   private _change:BehaviorSubject<ResourceUpdate>;
 
+  private _logger = viwiLogger.getInstance().getLogger("media.Renderers");
+
   constructor(private service:Service) {
 
-    const netfluxRendererId = "d6ebfd90-d2c1-11e6-9376-df943f51f0d8";//uuid.v1();  // FIXED for now
     //const collections = service.resources.map<Collections>(resource => resource.name === "collections");
     //const initialCollection = collections.map( element => element.name === "default");
     let netfluxRenderer = new BehaviorSubject<RendererElement>({
       lastUpdate: Date.now(),
       propertiesChanged: [],
       data: {
-        uri: "/" + this.service.name.toLowerCase() + "/" + this.name.toLowerCase() + "/" + netfluxRendererId,
-        id: netfluxRendererId,
+        uri: "/" + this.service.name.toLowerCase() + "/" + this.name.toLowerCase() + "/" + Renderers.netfluxRendererId,
+        id: Renderers.netfluxRendererId,
         name: "Netflux",
         state: "idle",
         shuffle: "off",
@@ -44,10 +53,11 @@ class Renderers implements Resource {
     this._renderers.push(netfluxRenderer);
 
     ///add an actual renderer for playback
-    let stpdId = uuid.v1();
+    let stpdId = Renderers.stdpRendererId;
     let stpdRenderer = new BehaviorSubject<RendererElement>({
       lastUpdate: Date.now(),
       propertiesChanged: [],
+      player: new stupid(),
       data: {
         uri: "/" + this.service.name.toLowerCase() + "/" + this.name.toLowerCase() + "/" + stpdId,
         id: stpdId,
@@ -99,26 +109,95 @@ class Renderers implements Resource {
 
 
   private _interval:NodeJS.Timer; //@TODO has to become per-renderer
+  private _playerInterval:NodeJS.Timer;//@TODO has to become per-renderer
+
   updateElement(elementId:string, difference:any):ElementResponse {
     let element = (<BehaviorSubject<RendererElement>> this.getElement(elementId).data);
-    let renderer:RendererObject = element.getValue().data;
+    var renderer:RendererObject = element.getValue().data;
     let propertiesChanged:string[]=[];
+
     if (difference.hasOwnProperty("state")) {
       renderer.state = difference.state;
       if (difference.state === "play") {
-        const speed = 1000;
-        this._interval = setInterval(() => {
-          renderer.offset = renderer.hasOwnProperty("offset") ? renderer.offset + speed : 0;
-          element.next(
-            {
-              lastUpdate: Date.now(),
-              propertiesChanged: ["offset"],
-              data: renderer
-            });
-        }, speed);
+        if (renderer.id === Renderers.netfluxRendererId)
+        {
+          const speed = 1000;
+          this._interval = setInterval(() => {
+            renderer.offset = renderer.hasOwnProperty("offset") ? renderer.offset + speed : 0;
+            element.next(
+              {
+                lastUpdate: Date.now(),
+                propertiesChanged: ["offset"],
+                data: renderer
+              });
+          }, speed);
+        } else if(renderer.name === "stpd" && element.getValue().player) {
+          let path = require("path");
+          let player = element.getValue().player; //might be undefinied, be aware
+
+
+          var onPlay = ()=> {
+            this._playerInterval = setInterval(() => { //@TODO: rather listen to player events..
+              renderer.offset = player.getOffset();
+              if(player.state) {
+                element.next({
+                  lastUpdate: Date.now(),
+                  propertiesChanged: ["offset"],
+                  data: renderer,
+                  player: player
+                });
+              }
+            }, 250);
+          };
+
+          var onPlayError = (err:Error)=>{
+            console.log("Player.play():", err.message);
+          };
+
+          switch (player.state) {
+            case "paus":
+              player.resume().then(onPlay, onPlayError);
+              break;
+            case "stop":
+              player.play(path.join(__dirname, '../../../samples/dimitriVegas.mp3')).then(onPlay, onPlayError);
+              break;
+            default:
+              return {status: "error", code: 500, error: new Error("unknown player state:" + player.state)}
+          }
+        }
       }
       else {
-        clearInterval(this._interval);
+        switch (renderer.id) {
+          case Renderers.netfluxRendererId:
+             clearInterval(this._interval);
+            break;
+          case Renderers.stdpRendererId:
+            let player = element.getValue().player; //might be undefinied, be aware
+            console.log(player);
+            switch (difference.state) {
+              case "pause":
+                if(player.state === "play") {
+                  player.pause();
+                  clearInterval(this._playerInterval); //@TODO: rather listen to player events..
+                } else {
+                  return {status: "error", error: new Error("Renderer not playing"), code: 500};
+                }
+                break;
+              case "stop":
+                if(player.state === "play") {
+                  player.stop();
+                  clearInterval(this._playerInterval); //@TODO: rather listen to player events..
+                } else {
+                  return {status: "error", error: new Error("Renderer not playing"), code: 500};
+                }
+                break;
+              default:
+                return {status: "error", error: new Error("Renderer state not supported"), code: 400};
+            }
+            break;
+          default:
+            return {status: "error", error: new Error("Renderer not found"), code: 404};
+        }
       }
       propertiesChanged.push("state");
     }
@@ -152,6 +231,10 @@ interface CollectionElement extends Element {
 class Collections implements Resource {
   private _collections:BehaviorSubject<CollectionElement>[] = [];
   private _change:BehaviorSubject<ResourceUpdate>;
+  private _medialibrary:Medialibrary;
+  private _tracks:Tracks;
+
+  private _logger = viwiLogger.getInstance().getLogger("media.Collections");
 
   constructor(private service:Service) {
 
@@ -169,7 +252,8 @@ class Collections implements Resource {
     });
     this._collections.push(initialCollection);
     this._change = new BehaviorSubject(<ResourceUpdate>{lastUpdate: Date.now(), action: 'init'});
-
+    this._medialibrary = new Medialibrary();
+    this._tracks = <Tracks>this._medialibrary.getResource("Tracks");
   }
 
   get name():string {
@@ -188,6 +272,39 @@ class Collections implements Resource {
     return this._change;
   }
 
+  private _setItems(itemuris:string[]):ItemObject[] | ElementResponse {
+    let _items:ItemObject[] = [];
+    let regex = /[\w\/\:]*([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fAF]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12})$/
+    let errors:string[] = [];
+    if(this._tracks) {
+      for (var index in itemuris) {
+        let uri = itemuris[index];
+        let match = uri.match(regex);
+        if (match !== null) {
+          let id = match[1];
+          if (!id) {
+            errors.push("Id " + id + "not valid from uri: " + uri);
+            console.log(errors)
+          } else {
+            let track = this._tracks.getElement(id);
+            if (track && track.data) {
+              _items.push(track.data.getValue().data);
+            }
+            else {
+              errors.push("Track " + id + " can not be found");
+            }
+          }
+        }
+      }
+      if (errors.length != 0) {
+        return {status: "error", error: new Error(errors.join(",")), code: 400};
+      } 
+    } else {
+      return {status: "error", error: new Error("No tracks loaded..."), code: 500};
+    }
+    return _items;
+  }
+
   getElement(elementId:string):ElementResponse {
     // find the element requested by the client
     return {
@@ -197,13 +314,23 @@ class Collections implements Resource {
     })};
   };
 
-  createElement(state:{name:string}):ElementResponse{
+  createElement(state:any):ElementResponse{
     if (!state.name) return {
       status: "error",
+      error: new Error('providing a name is mandatory'),
       code: StatusCode.INTERNAL_SERVER_ERROR
     };
     const collectionId = uuid.v1();
-    let initialCollection = new BehaviorSubject<CollectionElement>(
+    let items:any = [];
+
+    /** add items given with the query */
+    if(state.hasOwnProperty('items')) {
+      items = this._setItems(state.items);
+      if (!Array.isArray(items)) return items;
+    }
+
+    /** build the actual media collection and add it to the collections*/
+    let newCollection = new BehaviorSubject<CollectionElement>(
       {
         lastUpdate: Date.now(),
         propertiesChanged: [],
@@ -211,14 +338,34 @@ class Collections implements Resource {
         uri: "/" + this.service.name.toLowerCase() + "/" + this.name.toLowerCase() + "/" + collectionId,
         id: collectionId,
         name: state.name,
-        items: []
+        items: items
       }
     });
-    this._collections.push(initialCollection);
+    this._collections.push(newCollection);
+
+    /** publish a resource change */
     this._change.next({lastUpdate: Date.now(), action: "add"});
-    return {status:"ok", data: initialCollection};
+
+    /** return success */
+    return {status:"ok", data: newCollection};
   };
 
+  updateElement(elementId:string, difference:any):ElementResponse {
+    let element = this.getElement(elementId).data;
+    var collection:CollectionElement = element.getValue();
+    let propertiesChanged:string[]=[];
+
+    if(difference.hasOwnProperty('items')) {
+      let newItems = this._setItems(difference.items);
+      if (!Array.isArray(newItems)) return newItems;
+      collection.data.items = newItems;
+      collection.lastUpdate = Date.now();
+      collection.propertiesChanged = ['items'];
+      element.next(collection);
+      return {status: "ok"};
+    }
+    return {status: "error", code: 400, message: "No actions to take.."};
+  }
 
   deleteElement(elementId:string):ElementResponse {
     let idx = this._collections.findIndex((element:BehaviorSubject<CollectionElement>, index:number) => {