+++ src/plugins/medialibrary/index.spec.ts
@@ -1,0 +1,8 @@
+import * as Medialibrary from "./index";
+import { BehaviorSubject } from "@reactivex/rxjs";
+import { Element, ElementResponse, CollectionResponse} from "../viwiPlugin";
+
+
+describe("Service", () => {
+});
+